refactor(suite): move earn dashboard and voting delegation modules
