+++ packages/suite/src/components/earn/dashboard/EarnDashboard.tsx
