+++ packages/suite/src/components/earn/dashboard/common/EarnAccountCellDetails.tsx
