+++ packages/suite/src/components/earn/dashboard/common/EarnDashboardSection.tsx
