+++ packages/suite/src/components/earn/dashboard/common/EarnDashboardTableHeader.tsx
