+++ packages/suite/src/components/earn/dashboard/common/EarnInactiveNetworkOpportunity.tsx
