+++ packages/suite/src/components/earn/dashboard/common/EarnRewardsAmount.tsx
