+++ packages/suite/src/components/earn/dashboard/common/types.ts
