+++ packages/suite/src/components/earn/dashboard/staking/EarnStakingAccountRow.tsx
