+++ packages/suite/src/components/earn/dashboard/staking/EarnStakingActivateRow.tsx
