+++ packages/suite/src/components/earn/dashboard/staking/hooks/useCryptoCurrencyRate.tsx
