+++ packages/suite/src/components/earn/dashboard/utils/earnDashboardBadgeUtils.ts
