+++ packages/suite/src/components/earn/dashboard/yield/EarnYieldAccountOpportunity.tsx
