+++ packages/suite/src/components/earn/dashboard/yield/EarnYieldInactiveVaultOpportunity.tsx
