+++ packages/suite/src/components/earn/dashboard/yield/hooks/useAllYieldOpportunities.ts
