+++ packages/suite/src/components/earn/modals/shared/VotingDelegations/VotingDelegations.tsx
