+++ packages/suite/src/components/earn/modals/shared/VotingDelegations/VotingDelegationsOptions.tsx
